+++ guillotina_elasticsearch/tests/conftest.py
@@ -1,4 +1,5 @@
 pytest_plugins = [
+    'pytest_docker_fixtures',
     'guillotina.tests.fixtures',
     'guillotina_elasticsearch.tests.fixtures',
 ]